Varios endpoints y componentes de registrar postulante desarrollados
