+++ backend/app/Models/RegistroTutor.php
@@ -4,8 +4,9 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\Pivot;
 
-class RegistroTutor extends Model
+class RegistroTutor extends Pivot
 {
     use HasFactory;
