+++ frontend/service/registros.api.js
@@ -1,5 +1,15 @@
 import axiosInstance from "../helpers/axios-config";
 
+export const getRegistroByCI = async ($idOlimpiada, $ci) => {
+    try {
+        const response = await axiosInstance.get(`/registro/olimpiada/${$idOlimpiada}/postulante/${$ci}`);
+        return response.data;
+    } catch (error) {
+        console.error("Error fetching registro por ci:", error);
+        throw error;
+    }
+}
+
 export const enviarRegistrosLote = async (data) => {
     try {
         const response = await axiosInstance.post("/registro_lote", data);
